+++ packages/ai/core/generate-text/generate-text-result.ts
@@ -8,6 +8,7 @@ import { Source } from '../types/language-model';
 import { LanguageModelRequestMetadata } from '../types/language-model-request-metadata';
 import { LanguageModelResponseMetadata } from '../types/language-model-response-metadata';
 import { LanguageModelUsage } from '../types/usage';
+import { GeneratedFile } from './generated-file';
 import { ReasoningDetail } from './reasoning-detail';
 import { ResponseMessage, StepResult } from './step-result';
 import { ToolCallArray } from './tool-call';
@@ -30,6 +31,11 @@ has only generated text.
    */
   // TODO v5: rename to `reasoningText`
   readonly reasoning: string | undefined;
+
+  /**
+The files that were generated. Empty array if no files were generated.
+     */
+  readonly files: Array<GeneratedFile>;
 
   /**
 The full reasoning that the model has generated.