+++ packages/ai/core/generate-text/run-tools-transformation.ts
@@ -15,6 +15,7 @@ import {
 } from '../types';
 import { Source } from '../types/language-model';
 import { calculateLanguageModelUsage } from '../types/usage';
+import { DefaultGeneratedFileWithType, GeneratedFile } from './generated-file';
 import { parseToolCall } from './parse-tool-call';
 import { ToolCallUnion } from './tool-call';
 import { ToolCallRepairFunction } from './tool-call-repair';
@@ -38,6 +39,9 @@ export type SingleRequestTextStreamPart<TOOLS extends ToolSet> =
       type: 'redacted-reasoning';
       data: string;
     }
+  | ({
+      type: 'file';
+    } & GeneratedFile)
   | {
       type: 'source';
       source: Source;
@@ -158,6 +162,16 @@ export function runToolsTransformation<TOOLS extends ToolSet>({
         case 'response-metadata':
         case 'error': {
           controller.enqueue(chunk);
+          break;
+        }
+
+        case 'file': {
+          controller.enqueue(
+            new DefaultGeneratedFileWithType({
+              data: chunk.data,
+              mimeType: chunk.mimeType,
+            }),
+          );
           break;
         }