+++ packages/ai/core/generate-text/to-response-messages.ts
@@ -1,4 +1,5 @@
 import { ToolResultPart } from '../prompt';
+import { GeneratedFile } from './generated-file';
 import { ReasoningDetail } from './reasoning-detail';
 import { ResponseMessage } from './step-result';
 import { ToolCallArray } from './tool-call';
@@ -10,6 +11,7 @@ Converts the result of a `generateText` call to a list of response messages.
  */
 export function toResponseMessages<TOOLS extends ToolSet>({
   text = '',
+  files,
   reasoning,
   tools,
   toolCalls,
@@ -18,6 +20,7 @@ export function toResponseMessages<TOOLS extends ToolSet>({
   generateMessageId,
 }: {
   text: string | undefined;
+  files: Array<GeneratedFile>;
   reasoning: Array<ReasoningDetail>;
   tools: TOOLS;
   toolCalls: ToolCallArray<TOOLS>;
@@ -35,7 +38,13 @@ export function toResponseMessages<TOOLS extends ToolSet>({
           ? { ...part, type: 'reasoning' as const }
           : { ...part, type: 'redacted-reasoning' as const },
       ),
-      { type: 'text', text },
+      // TODO language model v2: switch to order response content (instead of type-based ordering)
+      ...files.map(file => ({
+        type: 'file' as const,
+        data: file.base64,
+        mimeType: file.mimeType,
+      })),
+      { type: 'text' as const, text },
       ...toolCalls,
     ],
     id: messageId,