+++ packages/openai/src/convert-to-openai-chat-messages.ts
@@ -58,7 +58,7 @@ export function convertToOpenAIChatMessages({
 
         messages.push({
           role: 'user',
-          content: content.map(part => {
+          content: content.map((part, index) => {
             switch (part.type) {
               case 'text': {
                 return { type: 'text', text: part.text };
@@ -101,7 +101,15 @@ export function convertToOpenAIChatMessages({
                       input_audio: { data: part.data, format: 'mp3' },
                     };
                   }
-
+                  case 'application/pdf': {
+                    return {
+                      type: 'file',
+                      file: {
+                        filename: part.filename ?? `part-${index}.pdf`,
+                        file_data: `data:application/pdf;base64,${part.data}`,
+                      },
+                    };
+                  }
                   default: {
                     throw new UnsupportedFunctionalityError({
                       functionality: `File content part type ${part.mimeType} in user messages`,
@@ -130,10 +138,6 @@ export function convertToOpenAIChatMessages({
               text += part.text;
               break;
             }
-            case 'redacted-reasoning':
-            case 'reasoning': {
-              break; // ignored
-            }
             case 'tool-call': {
               toolCalls.push({
                 id: part.toolCallId,
@@ -144,10 +148,6 @@ export function convertToOpenAIChatMessages({
                 },
               });
               break;
-            }
-            default: {
-              const _exhaustiveCheck: never = part;
-              throw new Error(`Unsupported part: ${_exhaustiveCheck}`);
             }
           }
         }