+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,62 @@
 
 ### Patch Changes
 
+- Updated dependencies [f10f0fa]
+  - ai@4.2.2
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [b796152]
+- Updated dependencies [06361d6]
+  - ai@4.2.1
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5bc638d]
+  - ai@4.2.0
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [5d0fc29]
+  - ai@4.1.66
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [16c444f]
+  - ai@4.1.65
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@4.1.64
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [0bd5bc6]
+  - ai@4.1.63
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [c9ed3c4]
+  - ai@4.1.62
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [2e1101a]
   - ai@4.1.61