+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/provider
 
+## 1.1.0
+
+### Minor Changes
+
+- 5bc638d: AI SDK 4.2
+
+## 1.0.12
+
+### Patch Changes
+
+- 0bd5bc6: feat (ai): support model-generated files
+
 ## 1.0.11
 
 ### Patch Changes